# determineDataLimits optimizations