--- a/src/controllers/controller.bar.js
+++ b/src/controllers/controller.bar.js
@@ -384,7 +384,7 @@ module.exports = DatasetController.extend({
 			value = custom.barStart;
 			length = custom.barEnd - custom.barStart;
 			// bars crossing origin are not stacked
-			if (value !== 0 && Math.sign(value) !== Math.sign(custom.barEnd)) {
+			if (value !== 0 && helpers.sign(value) !== helpers.sign(custom.barEnd)) {
 				start = 0;
 			}
 			start += value;