--- a/src/core/core.controller.js
+++ b/src/core/core.controller.js
@@ -391,6 +391,10 @@ helpers.extend(Chart.prototype, /** @lends Chart */ {
 				scales[scale.id] = scale;
 			}
 
+			// parse min/max value, so we can properly determine min/max for other scales
+			scale._parsedMin = scale._parse(scale.options.ticks.min);
+			scale._parsedMax = scale._parse(scale.options.ticks.max);
+
 			// TODO(SB): I think we should be able to remove this custom case (options.scale)
 			// and consider it as a regular scale part of the "scales"" map only! This would
 			// make the logic easier and remove some useless? custom code.