--- a/src/core/core.datasetController.js
+++ b/src/core/core.datasetController.js
@@ -3,7 +3,6 @@
 var helpers = require('../helpers/index');
 
 var resolve = helpers.options.resolve;
-var isNullOrUndef = helpers.isNullOrUndef;
 
 var arrayEvents = ['push', 'pop', 'shift', 'splice', 'unshift'];
 
@@ -187,17 +186,12 @@ function getFirstScaleId(chart, axis) {
 }
 
 function getUserBounds(scale) {
-	var min = scale._userMin;
-	var max = scale._userMax;
-	if (isNullOrUndef(min) || isNaN(min)) {
-		min = Number.NEGATIVE_INFINITY;
-	}
-	if (isNullOrUndef(max) || isNaN(max)) {
-		max = Number.POSITIVE_INFINITY;
-	}
-	return {min, max};
+	var {min, max, minDefined, maxDefined} = scale._getUserBounds();
+	return {
+		min: minDefined ? min : Number.NEGATIVE_INFINITY,
+		max: maxDefined ? max : Number.POSITIVE_INFINITY
+	};
 }
-
 // Base class for all dataset controllers (line, bar, etc)
 var DatasetController = function(chart, datasetIndex) {
 	this.initialize(chart, datasetIndex);
@@ -638,7 +632,7 @@ helpers.extend(DatasetController.prototype, {
 		var stacked = canStack && meta._stacked;
 		var indices = getSortedDatasetIndices(chart, true);
 		var otherScale = this._getOtherScale(scale);
-		var otherScaleBounds = getUserBounds(otherScale);
+		var {otherMin, otherMax} = getUserBounds(otherScale);
 		var i, item, value, parsed, stack, min, minPositive, otherValue;
 
 		min = minPositive = Number.POSITIVE_INFINITY;
@@ -649,7 +643,7 @@ helpers.extend(DatasetController.prototype, {
 			value = parsed[scale.id];
 			otherValue = parsed[otherScale.id];
 			if (item.hidden || isNaN(value) ||
-				otherScaleBounds.min > otherValue || otherScaleBounds.max < otherValue) {
+				otherMin > otherValue || otherMax < otherValue) {
 				continue;
 			}
 			if (stacked) {