… test/fixtures/controller.line/clip/default-x-max.png
… test/fixtures/controller.line/clip/default-x.png
… test/fixtures/controller.line/clip/default-y-max.png
… test/fixtures/controller.line/clip/default-y.png
… test/specs/controller.bar.tests.js
… test/specs/core.scale.tests.js
… test/specs/scale.logarithmic.tests.js