* `Math.sign` replaced with `helpers.sign`, for IE compatibility
* `ticks.min` / `ticks.max` parsed at scale creation for early access.
* `_getOtherScale` internal helper
* `controller._getMinMax`: skip values limited out of chart by other scale
* `scale._getMinMax`: skip loops if valid `min` & `max` are defined
* `scale.category`: use `_parsedMin` / `_parsedMax` in `determineDataLimits`
* `scale.logarithmic`: rename `nonNegativeOrDefault` to `positiveOrDefault`, better minimization

Resolves: #4131
[Pen](https://codepen.io/kurkle/pen/rNNdwJN?editors=1010)