@kurkle 2019-11-06
Rebased + added some optimizations to time scale. Removal of duplicate `arrayUnique` when ticks come from only one source (data / labels) in `getAllTimestamps` is noticeable (somewhere between 5-10% in uPlot bench).