@kurkle 2019-11-07
New uPlot benchmarks:
```
master - series: ~850ms
master - linear: ~650ms
2.9.2 - series: ~700ms
2.9.2 - linear: ~550ms
pr - series: ~700ms
pr - linear: ~450ms
```