@kurkle 2019-11-12
The improved (only visible data is accounted for) min/max determination changed scale ranges on some of those `clip` fixtures. Updated the images.